make Extension have the same Structure - Resources
