+++ www/front_src/src/Extensions/Listing/ExtensionDetailsDialog/LoadingSkeleton.tsx
